``` refactor: extract sidebar content logic into separate component
- Created SidebarContentWrapper component to handle sidebar menu rendering logic
- Moved mergeProjectsAndMenu function and SidebarProjects logic from AppSidebar to new component
- Removed unused imports from app-sidebar.tsx (icons, routing, permissions, constants)
- Simplified AppSidebar to focus on layout structure and pass props to SidebarContentWrapper
- Maintained existing functionality while improving code organization and separation
